Renamed contentable trait
