+++ src/Models/Contentable/Link.php
@@ -3,7 +3,7 @@
 namespace Illegal\Linky\Models\Contentable;
 
 use Illegal\LaravelUtils\Contracts\HasPrefix;
-use Illegal\Linky\Traits\Contentable;
+use Illegal\Linky\Contracts\Contentable;
 use Illuminate\Database\Eloquent\Model;
 
 class Link extends Model
